Add role-based management

+++ ToshokanApp/src/ToshokanApp.Presentation/Controllers/BookCommentController.cs
@@ -1,3 +1,4 @@
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Mvc;
 using ToshokanApp.Core.Models;
 using ToshokanApp.Core.Services;
@@ -18,31 +19,30 @@ public class BookCommentController : Controller
     {
         var bookComments = await this.bookCommentService.GetAllAsync();
 
-        return base.View(model: bookComments);
-    }
-
-    public IActionResult Add()
-    {
-        return View();
+        return base.View(bookComments);
     }
 
     [HttpPost]
-    [Route("[controller]")]
+    [Route("api/[controller]")]
     public async Task<IActionResult> Add(BookComment bookComment)
     {   
         if(ModelState.IsValid){
             await bookCommentService.AddAsync(bookComment);
-            return RedirectToAction(nameof(Index));
+            return Created();
         }
-        return View(bookComment);
+        return Forbid();
     }
 
     [HttpPost]
-    [Route("[controller]/{bookId}")]
+    [Route("api/[controller]/{bookId}")]
     public async Task<IActionResult> Delete(Guid bookId)
     {
-        await this.bookCommentService.DeleteAsync(bookId);
+        if (ModelState.IsValid)
+        {
+            await this.bookCommentService.DeleteAsync(bookId);
+            return base.Ok();
+        }
 
-        return base.RedirectToAction(nameof(Index));
+        return Forbid();
     }
 }
